# swagger/v1/swagger.yaml
swagger: '2.0'
info:
  description: >-
    SMX api description.
  version: 1.0.0
  title: MoneyApi
  termsOfService: ''
  contact:
    email: 
host: 107.22.146.137
basePath: /api/v1
...
schemes:
  - http
securityDefinitions:
  X-API-APIKEY:
    name: X-API-APIKEY
    in: header
    description: logged in user api key
    type: apiKey
paths:
...
  /pictures/profile_picture:
    get:
      tags:
        - profile
      summary: Get the profile picture
      description: The endpoint will get the picture of profile in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: query
          name: auth_token
          description: Auth token of the user
          required: true
          schema:
            $ref: '#/definitions/AuthToken'
      responses:
        '200':
          description: Profile picture
          schema:
            $ref: '#/definitions/PictureResponse'
        '401':
          $ref: '#/responses/401'
        '404':
          $ref: '#/responses/404'
        '500':
          $ref: '#/responses/500'
  /pictures/update_picture:
    patch:
      tags:
        - profile
      summary: Update a profile picture
      description: The endpoint will update a profile picture in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: formData
          name: auth_token
          description: Auth token of the user
          required: true
          schema:
            $ref: '#/definitions/AuthToken'
        - in: formData
          name: user[:avatar]
          description: Avatar file to upload
          required: true
          type: file
      responses:
        '201':
          description: Successful picture uploaded
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          $ref: '#/responses/400'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
...
definitions:
  AuthToken:
    type: object
    required:
      - auth_token
    properties:
      auth_token:
        type: string
        description: user api key
...
  PictureResponse:
    allOf:
      - type: object
        properties:
          updated_at:
            type: string
            description: Picture update datetime
            format: date-time
            example: '2021-07-15T18:08:36.189Z'
          avatar:
            type: string
            description: S3 bucket url of the profile picture
...
  Error:
    type: object
    required:
      - errors
    properties:
      errors:
        type: array
        items:
          type: string
          description: error message
  SuccessResponse:
    type: object
    required:
      - status
    properties:
      status:
        type: string
        description: Sucecess message
...
responses:
  '400':
    description: Input invalid or Bad Request
    schema:
      $ref: '#/definitions/Error'
  '401':
    description: Token Invlaid
    schema:
      $ref: '#/definitions/Error'
  '404':
    description: Resource not found
    schema:
      $ref: '#/definitions/Error'
  '500':
    description: Internal Server error
    schema:
      $ref: '#/definitions/Error'